# backend/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Luma Local Media API
  version: 0.1.0
servers:
  - url: /
paths:
...
  /api/v1/sources:
    get:
      summary: 查询媒体源列表
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: 不包含真实根路径的媒体源列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Source" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: 创建本地媒体源
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSource" }
      responses:
        "201":
          description: 媒体源已创建
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "403": { $ref: "#/components/responses/ForbiddenPath" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/sources/{id}:
    parameters:
      - $ref: "#/components/parameters/SourceID"
    patch:
      summary: 更新或重新启用媒体源
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSource" }
      responses:
        "200":
          description: 更新后的媒体源
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      summary: 软禁用媒体源并保留全部索引
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: 媒体源已禁用 }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/admin/users/{id}/sources:
    get:
      summary: 查询成员可访问的媒体源（仅管理员）
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: 来源 ID 列表
          content:
            application/json:
              schema:
                type: object
                required: [source_ids]
                properties:
                  source_ids: { type: array, items: { type: string } }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/admin/users/{id}/sources/{sourceId}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
      - { in: path, name: sourceId, required: true, schema: { type: string } }
    put:
      summary: 授予媒体源访问权（仅管理员）
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: 已授权 }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      summary: 移除媒体源访问权（仅管理员）
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: 已移除 }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
    CreateSource:
      type: object
      additionalProperties: false
      required: [name, root_path]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        root_path: { type: string, minLength: 1 }
        library_kind: { type: string, enum: [personal, movies, tv], default: personal }
    UpdateSource:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        root_path: { type: string, minLength: 1 }
        enabled: { type: boolean }
        library_kind: { type: string, enum: [personal, movies, tv] }
    Source:
      type: object
      required: [id, name, type, library_kind, enabled, status, created_at, updated_at]
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string, const: local }
        library_kind: { type: string, enum: [personal, movies, tv] }
        enabled: { type: boolean }
        status: { type: string, enum: [online, offline, degraded, disabled] }
        last_scan_id: { type: string }
        last_seen_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
  parameters:
    SourceID:
      in: path
      name: id
      required: true
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: API Token 缺失或无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InvalidRequest:
      description: 请求参数、请求体或业务输入无效
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ForbiddenPath:
      description: 媒体源路径不在白名单中
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: 当前身份没有执行该管理操作的权限
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Conflict:
      description: 资源状态或唯一性冲突
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }